Comment lines feature, remove docopt

+++ tests/nimgentest.nims
@@ -1,7 +1,18 @@
+import distros
 import ospaths
 import strutils
 
-for comp in @["nimbass", "nimfuzz", "nimssl", "nimssh2"]:
-    withDir(".."/comp):
-        exec "nimble install -y"
-        exec "nimble test"
+var comps = @["libsvm", "nimbass", "nimfuzz", "nimssl", "nimssh2"]
+if detectOs(Windows):
+    comps.add("nimkerberos")
+
+for comp in comps:
+    if dirExists(".."/comp):
+        exec "nimble uninstall -y " & comp
+        withDir(".."/comp):
+            rmDir(comp)
+            exec "nimble install -y"
+            exec "nimble test"
+
+            exec "nimble install -y"
+            exec "nimble test"
